[Data Explorer] Timestamp overflow bugfix

+++ libraries/data_explorer/src/main/java/com/tokopedia/data_explorer/db_explorer/data/sources/cursor/shared/CursorSource.kt
@@ -51,7 +51,7 @@ internal open class CursorSource {
                 )
                 FieldType.INTEGER -> Field(
                     type = type,
-                    text = cursor.getInt(column).toString()
+                    text = cursor.getLong(column).toString()
                 )
                 FieldType.FLOAT -> Field(
                     type = type,
